rename folder, begin delete function
